chore: Move functions to upload middleware in core
